mmm_model/requirements.txt
numpy==1.26.0
scipy==1.12.0
jaxlib==0.4.18
numpyro==0.11.0
lightweight-mmm==0.1.9
pandas==1.5.3
scikit-learn==1.2.2
google-cloud-storage==2.10.0
google-cloud-bigquery==3.10.0
google-auth==2.23.4
db-dtypes==1.0.0
